requirements_desktop.txt
gradio
streamlit
numpy
pillow
pillow-heif
matplotlib
lameenc
pyinstaller
pywebview
